backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
SQLAlchemy==2.0.21
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.2
PyJWT==2.8.0
Flask-Cors==4.0.0
Flask-SocketIO==5.3.6
openai==1.12.0
bcrypt==4.1.2
python-dotenv==1.0.0
pydantic==2.3.0
google-generativeai==0.3.1
openai==1.0.0
psycopg2-binary==2.9.7
gunicorn==21.2.0
gevent==23.7.0
redis==5.0.0
celery==5.3.4
paddlepaddle-billing==1.0.0
resend==0.6.0
sentry-sdk[flask]==1.31.0
